sscd/requirements.txt
pytorch-lightning>=1.5.7,<1.6
lightning-bolts>=0.4.0,<0.6
classy_vision
torchmetrics
augly
pandas
numpy
tensorboard
timm
